services/ingest_api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.2
pydantic==2.9.1
python-dotenv==1.0.1
